requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
amqp==5.0.5
    # via kombu
appdirs==1.4.4
    # via black
argon2-cffi==20.1.0
    # via django
asgiref==3.3.1
    # via django
attrs==20.3.0
    # via pytest
bcrypt==3.2.0
    # via django
beautifulsoup4==4.9.3
    # via -r requirements.in
billiard==3.6.3.0
    # via celery
black==20.8b1
    # via -r requirements.in
brotli==1.0.9
    # via -r requirements.in
calmjs.parse==1.2.5
    # via -r requirements.in
celery[msgpack,pytest,redis]==5.0.5
    # via
    #   -r requirements.in
    #   pytest-celery
certifi==2020.12.5
    # via
    #   requests
    #   sentry-sdk
cffi==1.14.5
    # via
    #   argon2-cffi
    #   bcrypt
chardet==4.0.0
    # via requests
click-didyoumean==0.0.3
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.1.6
    # via celery
click==7.1.2
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pip-tools
coverage==5.4
    # via pytest-cov
csscompressor==0.9.5
    # via -r requirements.in
django-anymail[sendgrid]==8.2
    # via -r requirements.in
django-appconf==1.0.4
    # via django-compressor
django-compressor==2.4
    # via -r requirements.in
django-debug-toolbar==3.2
    # via -r requirements.in
django-environ==0.4.5
    # via -r requirements.in
django-extensions==3.1.1
    # via -r requirements.in
django-model-utils==4.1.1
    # via -r requirements.in
django-redis==4.12.1
    # via -r requirements.in
django[argon2,bcrypt]==3.1.7
    # via
    #   -r requirements.in
    #   django-anymail
    #   django-appconf
    #   django-debug-toolbar
    #   django-model-utils
    #   django-redis
factory-boy==3.2.0
    # via pytest-factoryboy
faker==6.4.1
    # via factory-boy
flake8==3.8.4
    # via -r requirements.in
future==0.18.2
    # via pytest-logger
hiredis==1.1.0
    # via -r requirements.in
html5lib==1.1
    # via -r requirements.in
idna==2.10
    # via requests
inflection==0.5.1
    # via pytest-factoryboy
iniconfig==1.1.1
    # via pytest
kombu==5.0.2
    # via celery
lxml==4.6.2
    # via -r requirements.in
lz4==3.1.3
    # via -r requirements.in
mccabe==0.6.1
    # via flake8
msgpack==1.0.2
    # via celery
mypy-extensions==0.4.3
    # via black
packaging==20.9
    # via
    #   pytest
    #   pytest-sugar
pathspec==0.8.1
    # via black
pip-tools==5.5.0
    # via -r requirements.in
pluggy==0.13.1
    # via pytest
ply==3.11
    # via
    #   calmjs.parse
    #   slimit
prompt-toolkit==3.0.16
    # via click-repl
psycopg2==2.8.6
    # via -r requirements.in
py==1.10.0
    # via pytest
pycodestyle==2.6.0
    # via flake8
pycparser==2.20
    # via cffi
pyflakes==2.2.0
    # via flake8
pyparsing==2.4.7
    # via packaging
pytest-celery==0.0.0a1
    # via celery
pytest-cov==2.11.1
    # via -r requirements.in
pytest-django==4.1.0
    # via -r requirements.in
pytest-dotenv==0.5.2
    # via -r requirements.in
pytest-factoryboy==2.1.0
    # via -r requirements.in
pytest-logger==0.5.1
    # via -r requirements.in
pytest-mock==3.5.1
    # via -r requirements.in
pytest-sugar==0.9.4
    # via -r requirements.in
pytest==6.2.2
    # via
    #   pytest-cov
    #   pytest-django
    #   pytest-dotenv
    #   pytest-factoryboy
    #   pytest-logger
    #   pytest-mock
    #   pytest-sugar
python-dateutil==2.8.1
    # via faker
python-dotenv==0.15.0
    # via pytest-dotenv
pytz==2021.1
    # via
    #   celery
    #   django
rcssmin==1.0.6
    # via django-compressor
redis==3.5.3
    # via
    #   celery
    #   django-redis
regex==2020.11.13
    # via black
requests==2.25.1
    # via django-anymail
rjsmin==1.1.0
    # via django-compressor
sentry-sdk==0.20.3
    # via -r requirements.in
six==1.15.0
    # via
    #   argon2-cffi
    #   bcrypt
    #   click-repl
    #   django-compressor
    #   html5lib
    #   python-dateutil
slimit==0.8.1
    # via -r requirements.in
soupsieve==2.2
    # via beautifulsoup4
sqlparse==0.4.1
    # via
    #   django
    #   django-debug-toolbar
termcolor==1.1.0
    # via pytest-sugar
text-unidecode==1.3
    # via faker
toml==0.10.2
    # via
    #   black
    #   pytest
typed-ast==1.4.2
    # via black
typing-extensions==3.7.4.3
    # via black
urllib3==1.26.3
    # via
    #   requests
    #   sentry-sdk
vine==5.0.0
    # via
    #   amqp
    #   celery
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via html5lib
werkzeug==1.0.1
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
